There's no need to insert spaces into urls anymore
Since bd7b180 (3.9.4) we use word-wrap: break-word on the message body,
which lets the browser break lines in the middle of words if it needs
to.  This CSS property is a lot cleaner than shotgunning in spaces or
&shy; or <wbr> every so often, and has great browser support:

    https://developer.mozilla.org/en/CSS/word-wrap#Browser_compatibility

If the more general application of word-wrap: break-word ever goes away
on message bodies, we should apply it to links within the message body.