Display embedded images in HTML messages inline with the text
Attached images with a Content-ID header referred to by an <img
src="cid:..."> tag in the HTML are now properly displayed as long as the
existing $ShowTransactionImages option is enabled.  If the option is
disabled, images are not inlined with the HTML.

This makes RT's message display even more like a real mail client's.