Warning avoidance during Create txn in TransactionBatch test
The Create transaction of $ticket causes a pending transaction for
TransactionBatch, which was not triggered until it passed out of scope
during global destruction.  As the applicable check tests ->Field eq,
and ->Field is undef on a Create transaction, this caused a warning.

Flush the transaction earlier, explicitly, and warnings-proof the
->Field check.