Restore database disconnection state after successful safe_run_child
RT::Util's safe_run_child sets its database handles to not disconnect
themselves if they are destroyed, before calling the provided function
which may fork and exec.  It explicitly re-enables those bits prior to
die'ing if the exec fails, to ensure that the database handle is torn
down correctly during the global destruction that would shortly ensue.

However, it fails to re-instate those bits after a _successful_ call.
This leaves the main database handle in a state where it does not tear
down the connection during global destruction.

This is particularly destructive in the case where:
  (a) RT uses PostgreSQL as its backend database
  (b) The database connection to PostgreSQL uses SSL, as is the default
      if the server supports it
  (c) The RT server is embedded into the Apache server using mod_perl
  (c) Apache has also loaded the SSL libraries for HTTPS support

This causes libcrypto.so to be used in two places in the Apache process,
by both Perl's binary PostgreSQL driver as well as core Apache's; they
thus share some internal state.  The lack of orderly teardown of the
SSL-enabled database connection causes corruption in the SSL engine's
internal state during the Apache shutdown process, which could lead to
segmentation faults in Apache.

Resolve this by explicitly re-instating the disconnect-on-destroy flags
after a successful safe_run_child.