Avoid shell interpolation when calling sendmailpipe
Use IPC::Open2 rather than three-arg open with "|-", to prevent shell
interpolation.  This allows not only fine control of the parameters to
the sendmail command, but also allows the STDOUT content to be dropped,
as was previously handled using a shell output redirect.  For backwards
compatability, use Text::ParseWords to split the configuration
parameters which are strings, rather than argument lists.

This prevents a vulnerability where specially-crafted emails could allow
code execution if VERP was enabled.

This resolves CVE-2011-2082.