Preserve page anchors across login tangents
Browsers carry over the anchor upon receiving a redirect response.  Take
advantage of this by appending any anchor the login page has to the
form's action.  When the login form is submitted successfully, a
redirect is issued and the browser carries over the anchor as intended
for the original page.

Confusion might arise if the login page ever shares an anchor name with
deeper pages.

Unfortunately there's no way to get at the anchor from the server side
at the time of tangent, otherwise we could simply stuff it into the
NextPage entry and be done with it.