Extract concept of many-to-many association and sorting
ObjectCustomFields contain the core concept of taking a set of objects
(custom fields) and applying them either globally or to a distinct
subset of other objects (queues).  Furthermore, it provides a way of
sorting those applications, including dealing with the complexities of
the interaction between global and queue-specific ordering.  This
concept is a useful one to abstract out into a RT::Record (and
RT::SearchBuilder) subclass, named ApplyAndSort.