Safely report fatal errors in children of safe_run_child in the parent
In the event of a fork, the RT logger object may have shared sockets
open, and as such may not be usable.  In order to ensure that the fatal
error in the child is not silently discarded, write it to a pipe which
is left open in the parent.  The contents of this pipe are read in the
parent, and re-thrown as a warning.