Cope with children of safe_run_child that don't immediately exit
By making the read from the pipe non-blocking, it deals gracefully with
the case where the child is still running, and thus has left the pipe
open.  This will eventually cause a zombie child, as it will not be
reaped, but does allow for safe_run_child to continue if its process
chooses to not wait() or waitpid() and the child survives.