Use sleep() and signals to remove a possible race condition between processes
It is possible that the parent could try to open the first file before
the child has had a chance to write it, which would lead to a spurious
test failure.  To avoid this, we establish a 3-second grace period
during which we sleep().  Since any signal ends a sleep() call early, we
will either be woken by a USR1 signal from the child when it is ready,
or we will simply carry on and check for existence after 3 seconds.  We
also install a trivial signal handler to allow us to skip the sleep()
entirely if the signal arrives before we would enter the sleep().

It is theoretically possible that on sufficiently heavily-loaded
systems, 3 seconds is insufficient; if the system is that heavily
loaded, however, you likely have other problems.