Ensure that I18N::LangTags::Detect only finds the "C" locale in tests
The I18N::LangTags::Detect package, which RT uses to auto-detect the
user's language preference in the browser, also looks at several
environment variables which are more commonly set on the command line.
This is only an issue during tests, as we expect the language to always
be English.

Explicitly always remove all of the environment variables which
I18N::LangTags::Detect looks at, and set LANG=C to provide a reference
point.