Make CheckIntegrity idempotent on a running install
Avoid reconnecting to the database if we have a handle and stop trying
to initialize the logger.  These changes make it safe to call as many
times as you'd like on a running instance without affecting the app.
Specifically, CheckIntegrity is now safely callable from places other
than server initialization.

Note that any calls to CheckIntegrity must now come _after_
InitLogging() is called.  In RT 4, CheckIntegrity is only called in
rt-server.