Test that RT::Users->WhoHaveRight doesn't pick up disabled groups
This test file currently fails.  The bug is noticeable in the owner
lists since SelectOwner is the only place which uses WhoHaveRight.

Fixing it requires checking Disabled on the Principals table joined
against Groups (not Users, which is already joined).  Alternatively, we
should be able to check Disabled on CachedGroupMembers as long as we
make sure the check always goes through CGM instead of sometimes
GroupMembers and sometimes CGM.