Force numerical ID sorting in local tickets
The sort order on links was based on Target or Base, not LocalTarget or
LocalBase, which meant that it was left to the database's infernal
devices as to how to sort those for local tickets (whose Target/Base is
NULL, but which have a LocalTarget/LocalBase).  On MySQL, for instance,
this led to lexical sorting: 1, 10, 11, 2, 3...

Sort by both LocalTarget/LocalBase as well as Target/Base.