Only redirect to the approvals view on GET requests
POST (and other) requests are likely to contain data that we don't want
to convert query parameters in a GET request.

All requests except GET are allowed through unmolested.  Hopefully after
the appropriate actions are taken, the page does a redirect using
MaybeRedirectForResults, which will generate a GET request that is
followed and subsequently caught by MaybeRedirectToApproval.

This behaviour is similar to the custom approval redirection logic in
Ticket/Display.html: it always does its redirect after any processing.

Ruslan++ for pointing out the flaw of redirecting POST requests.