Preserve the dispositions of attachments when redistributing mail
Use the original Content-Disposition if there is one, otherwise let
MIME::Entity default to inline.

MIME::Entity->build (used by ->attach) defaults to an inline disposition
if none is specified.  This had the affect of mutating all incoming
parts with an attachment disposition into inline content when RT
redistributed it to watchers.