Allow links to come before the tickets they reference
While the serialization process can _nearly_ guarantee that no object is
serialized before its dependencies, there are some times when it cannot
uphold this -- specifically, involving links to deleted tickets, for
instance.  In these cases, use the Postpone machinery (extended to allow
for ->URI) to update the link object later.