Remove HTML::TreeBuilder and HTML::FormatText from rt-mailgate's deps
They are still required by core RT.

HTTP::Response->error_as_HTML has always been a method that returns
static, predictable HTML based on the HTTP response code.  There is no
reason to parse it with a full blown HTML parser just so we can then
turn around and turn it _back_ into text.

Since we're in a stable series and who knows what might be parsing the
error message, take care to output the same format that was previously
returned.  We should consider trimming the error message for 4.2 to
remove the useless static text.