Apply the code changes from this branch to the mobile version as well
Even though the mobile interface does not currently provide a UI for
cloning tickets (the only interface is through Elements/ShowLinks, which
hardcodes /Ticket/Display.html), this keeps the two copies of the code
in synch.  This is so when we collapse these two codepaths back together
later, they will be more obviously the same code.