The incorrect additional / caused URI->rel to generate "///Articles/...."
WebURL should never have a explicitly / after it; while generally
harmless, it combines poorly with the use of URI->rel, which shortens it
to "///Articles/Articles/Article/Display.html", which means "the current
scheme, the host articles, and path /Articles/Article/Display.html"
which is absolutely incorrect.

Strip the extra / from the URI which arcticles' HREF was adding.