Instead of overwriting a database with initialdata, insert into an empty db
Attempting to determine which transactions and attached records should
be updated, as opposed made fresh, is a complicated business.  Instead
of fighting with possible inconsistency with the serialized dump and
"standard" initdb row ids, it is much simpler and safer to demand that
we insert into a completely empty database.

Rename --overwrite to --clone to reflect this change.  This also removes
a number of kludgite codepaths which attempted to merge if --overwrite
was passed.