Don't shift the datetime picker up if it would hit the bottom
Shifting it up weirdly covers up the input and sometimes runs into
z-index problems with the aileron menus.

The datetime picker now extends past the viewport bottom and will push
the page taller if it hits the page bottom.  This requires a short
scroll, but is preferable to inconsistent and awful looking alignment
otherwise.