email could be multipart already before calling AddAttachmentsFromHeaders
the bug's behavior is: when people select attachments and upload some too at
the same time, the uploaded attachments will leak to EmailRecord txn, which
results in duplicates for uploaded ones.

this is because in RT::Action::SendEmail::RecordOutgoingMailTransaction, we
filter top parts of email by checking their "RT-Attachment" headers, so we
can't make multipart again if the email is already multipart.