Serializer: Consistently and conveniently push collections onto the stack
PushCollections explicitly sets an order by id (ascending) so that it is
stable across databases and data.  This makes it easier to debug and
while it affects the order in which objects are walked, it shouldn't
make a net difference in the objects serialized.

For the sake of being explicit, FindAllRows is called on all collections
even though RT::DependencyWalker->Walk calls FindAllRows when it unrolls
collection objects on the stack.  If the default stack walking behaviour
ever changes, the meaning of PushCollections will not.