Enable LastUpdated on Scrips by not walking around RT::Record->_Set
75ad51c, as part of a "security audit," moved RT::Scrip's ->_Set method
to call ->__Set, rather than ->SUPER::_Set; this means that it skips the
functionality on RT::Record->_Set.  This skips the call to
->SetLastUpdated, causing all Scrips to never update their LastUpdated
ot LastUpdatedBy columns.  Less importantly, it also skips the code
which makes the return message explicit as to the values which changed;
while not strictly a feature in this case, it is at least consistent
with Templates and the like.

Revert to calling ->SUPER::_Set, so Scrips' LastUpdated columns are kept
up-to-date.