Cloning: Select pages of records via id > N rather than OFFSET
Under MySQL, using "LIMIT 2000 OFFSET 3000000" causes 3,002,000 rows to be
retrieved and then the first 3,000,000 are discarded.  Large collections
rapidly increase in query execution length when paged in this manner.
To avoid a massive performance hit when cloning, we simulate OFFSET with
an id > N restriction.  This takes advantage of the ordering by id and
lack of any other limits on id when cloning.