Catch warnings to STDERR when loading rt-server, for installer tests
The web installer tests introduced in f855c79 use the functionality
provided in 61dd5be, to load the 'sbin/rt-server' file, which sets up
the web installer if the database is missing.  Unfortunately, this file
prints to STDERR, which makes the test no longer warnings-clean.

Though this particular case prints to STDERR directly (rather than
calling warn), changing it would not improve matters, as the warnings
handler which Plack::Middleware::Test::StashWarnings installs only
happens at request time, much too late to catch the warning.

Explicitly catch the warning printed to STDERR by temporarily
redirecting STDERR to an in-memory filehandle, and recording its
contents as a warning by hand.  As this API was introduced in
Plack::Middleware::Test::StashWarnings 0.06, bump the dependency
accordingly.