Add basic HTTP_REFERER checking to prevent cross-site request forgery
RT has had a long-standing policy of accepting query parameters in both
POST and GET, and taking action based on them.  Unfortunately, this
makes RT susceptible to cross-site request forgery (CSRF), wherein the
browser is tricked into making a request on behalf of the user, which
includes the user's cached credentials, and the remote server runs an
malicious action with the user's identity.

Prevent this by enforcing that all requests to RT come with a "Referer"
(sic) header of RT itself.  Studies have shown that while the incidence
of suppressed or incorrect Referer headers is non-trivial over unsecured
HTTP (3% - 11%), it is very uncommon (0.05% - %0.22%) when the server is
secured using HTTPS[1].  As running RT without SSL protection already
presents a host of other vulnerabilities, including man-in-the-middle
attacks and password sniffing, we have judged that disabling CSRF
protections in HTTP-only deployments should be an acceptable solution.

Enforcing that all requests have such a header is entirely too
restrictive, as it prevents a large number of idempotent requests which
are not abusable, but provides the foundation of the CSRF protection.

This resolves CVE-2011-2085.

[1] http://seclab.stanford.edu/websec/csrf/csrf.pdf