Add a whitelist of idempotent request arguments
The previous commit introduced Referer enforcement; however, this
protection is entirely overzealous, as it prevents even the _first_
request to RT, and various other requests which would change no state on
the server.

Whitelist requests which contain no arguments, or which contain only
arguments which are assured to change no state on the server.  This
allows initial visits to http://rt.example.com/ to pass the Referer
check, as well as allowing existing links in the form of
http://rt.example.com/Ticket/Display.html?id=1 to function.