Redirect to an interstitial page on CSRF attacks, rather than denying
As some browsers never supply Referer headers, and to maintain better
integration with other sites which may link into RT, simply denying the
a potential CSRF request with no recourse is unfortunate, and
frustrating to the user.

Instead, store the request parameters in a unique key on the session,
and provide a page with a link which will extract them again and re-run
the request.  This allows intercepted requests to continue with only one
extra click.