Allow file uploads to persist across CSRF interstitial
While storing request parameters into the session is sufficient for most
requests, it does not cover the case of file uploads, when the request
arguments contain a filehandle, and thus cannot be cleanly serialized.

Deal with this, in the case of ticket attachments, by parsing and
storing the attachment in the CSRF session data using MakeMIMEEntity, as
the request would usually have done.  On the event of the CSRF token
being used, extract the parsed attachment and insert it into the
canonical session storage for attachments.

This does not cover the case of custom field file uploads, as the latter
are never stored in the session, and are much harder to isolate.