Add optional CSRF login protection
RT has historically allowed logins from any location by simply providing
'user' and 'pass' parameters, which were used to log the user in.  This
opens RT to "login CSRF," wherein a user is tricked into logging in
using the _attacker_'s credentials, which allows the attacker to later
examine their own account and possibly extract information about the
user's browsing history.

As RT does not track or display such information about the user, and the
ability to provide URLs which log a user into RT is central to many
workflows, explicitly whitelist 'user' and 'pass' parameters to RT for
CSRF purposes, if (and only if) they authenticate correctly.  If so, it
is also sufficient to whitelist the entire request, as it proves that
the requesting entity possesses the user's full credentials, and would
be able to take any action regardless.

This exception will not function if RT is configured to authenticate
against a source other than its internal database, such as any of those
provided by RT::Authen::ExternalAuth, or via the $ExternalAuth
configuration setting, as RT has no way of verifying the credentials
provided via the query parameters.

If login CSRF is deemed a potential issue locally, we provide a
configuration option ($RestrictLoginReferrer) to enable login CSRF
protection.