Allow REST requests to function regardless of Referer header
The REST interface, and existing clients to it, expect to be able to
make queries without supplying a Referer parameter -- which CSRF
protections should rightly deny.  While the simple solution would be to
include a Referer header in such requests, we cannot assume that all
REST clients can be so updated.

Create a secondary class of session, the REST session, which is allowed
to make Referer-less state-modifying requests, but not access the rest
of the UI.  We recognize such sessions by if their initial request is to
a REST URL; if so, we mark the session as a REST session.  For clients
with existing sessions which are not flagged as REST or not, we assign
them into the appropriate class based on the first request we observe;
this will allow existing clients with long-running sessions to continue
unmodified.

This also removes a SessionType check which became obsolete in 54f0f73,
and re-implements the functionality using the new REST flag on the
session.