Overhaul what CSS we allow in style attributes to be safer *and* more useful
The regex takes care to accomodate Outlook's idiosyncrasies and ensure
that the features of the rich text editor work.

This is still a large hammer.  If there's any forbidden CSS property
specified the entire style attribute will be stripped (including
otherwise allowed properties).  The fix for this is parsing the CSS
itself with something like CSS::Adaptor::Whitelist, but that's for
another day.