Load and Validate Custom Field Context Objects
We now require that you pass a ContextId in the AutoComplete URL and we
use that Id to load an object of the appropriate type and check that the
object is actually related to the Custom Field you're looking at (this
prevents you from using Queue A to look at valuds for Custom Field B
which is only applied to Queue C, or using a Class of id 8 to simulate a
Queue of id 8).

Since you might use the Context Object to gain rights to load a
CustomField, we can't just load a random Custom Field and ask "is this a
context object?" which is why this code loads a Custom Field as the
system user and then uses that privileged object to validate the context
object before dropping privileges and loading the CO and CF as the
current user.