Tell users and admins what Referrer we wanted
This introduces a normalizing method we could use elsewhere in the Web
code, as well as uses that code to hide the localhost->127.0.0.1
transformations.
It also adds to the error string so that you know what RT was expecting.
"RT's configured hostname" is the best we could do without explicitly
stating WebBaseURL in a user facing error message.