$r->path_info is not reliable; use the request_comp's path
$r->path_info is not populated under mod_perl with a non-empty WebPath,
due to the way HTML::Mason::ApacheHandler is implemented.  As what we
truely care about is if the component being requested is under the REST
directory, ->path_info can be replaced with ->request_comp->path, which
both simplifies the regex and brings it in line with the one used during
auth.