Set the refresh URL on ticket results to a CRSF-safe one
Unfortunately, browsers do not always provide a referrer when
redirecting to a page by way of a <meta http-equiv="refresh">.  As such,
automatic result refreshes trigger CSRF protections, as they request a
complex URL with many query parameters, and no referrer.

Work around this by generating a CSRF token which encodes the complete
set of query parameters, and redirecting to that.  An unfortunate but
unavoidable side effect of this is that the session is bloated by each
of these sets of store query parameters on each search result page that
has a refresh set.