convert text part to binary when we can not decode
If we couldn't find or guess charset or if charset
is not supported or failed to parse properly then
RT was dropping email.

Now instead of dropping we set content type to
'application/octet-stream' and save original value
to X-RT-Original-Content-Type header. In this
case users can download part and deal with it.

In most cases it would be related to spam, but
there are valid charsets that are not supported
by Encode module.