Dissalow undoumented numeric IDs in ticket create.
Currently, you can create a ticket (via the REST interface) with a Cc
of => [4], which will unexpectedly cause the ticket to be CC'd to the
entire group (all privilidged users is always ID #4).

This is obviously unexpected, and in a large multi-tenant shared RT situation,
all privilidged users seeing a specific ticket is really really bad..

If for some reason this API needs to stay, then I strongly think it should
refuse to create CCs for the default privilidged / everyone groups, as this
is never what you want, and always bad...

Also, the user interface is completely unable to handle tickets which have a
group CC'd in - they instead list a blank space, and checking the checkbox
doesn't work to remove the group. So even after we'd worked out what was
going on, we weren't able to undo it without manual database hacking.

If the numeric API _is_ meant to be supported, then it either needs to expand
group ID into a list of user IDs and add these, or group ID#s need to be
disallowed, or the interface needs to correctly support the removal of groups
from tickets.