Fix sending email with 'perl-script' mod_perl handler and 'sendmailpipe'
ef35fe5 switched to using IPC::Open2 to communicate with the sendmail
process.  Unfortunately, due to the way it is implemented, IPC::Open2
assumes that the filehandle named STDIN is always on file descriptor 0;
if it is not, the child process does not recieve the expected pipe on
its FD 0, and mail fails to be sent.  This is manifest either as a
SIGPIPE when the parent attempts to read from their end of the pipe, or
an abnormal exit code if the child exits (noting it has no input) before
that.

Apache's mod_perl, when running under the 'perl-script' handler, not
only TIEs STDIN and STDOUT, as noted in d80b1e4, but also moves them to
not be on FD 0 and FD 1.  While RT 4.0 has moved to requiring the
'modperl' handler (for more or less this exact reason), RT 3.8 is not
prepared for such.

In order to ensure that IPC::Open2 works on all deployment options,
locally override STDIN to be on file descriptor 0, regardless of what FD
0 is currently being used for.