Protect both FD 0 and FD 1 on mod_perl to avoid sendmailpipe errors
Under the 'perl-script' handler on mod_perl, STDIN and STDOUT are
effectively moved out of the way before the body of the request is run,
such that they always appear to be closed in the handler.
Unfortunately, the way this was done on mod_perl <= 2.0.4 was to
actually close them, while storing away the handle in a higher FD for
later restoration.  This means that, after the completion of the first
request, FD 0 and FD 1 were left available for later system calls to
obtain, as STDIN and STDOUT crept into higher-numbered FDs.

While we explicitly opened FD 1 during the first request to protect
against this sort of issue, no such protection was given to FD 0.  This
allowed it to be claimed, in the second request, by the socket to the
browser.  This causes failures in 'sendmailpipe' communication, as we
are unable to dup STDIN onto FD 0 prior to calling IPC::Open2 (see
b7a5a53).  The symptoms of this in Apache's error log varied, but
included:

  Undefined value assigned to typeglob at RT/Interface/Email.pm line 459

  Apache2::RequestIO::rflush: (103) Software caused connection abort

  (9)Bad file descriptor: core_output_filter: writing data to the network

Though these errors were only observed under the prefork MPM, there is
no known reason that they could not occur under the worker MPM as well.

Resolve this by ensuring that we always claim both FD 0 and FD 1 during
the first request.  In doing so, we replace the possibly-brittle call to
open(), where we can only hope we get the FD we want, with
IO::Handle->new_from_fd(), which allows us to specify the FD we want.

mod_perl 2.0.5 resolved this issue by never actually closing STDIN and
STOUT, but rather local'ing them to be undef; this ensures that no later
system calls will obtain them accidentally.  Forcing them open using
new_from_fd, as this commit does, does not hurt matters, however.