Serializer: Use the same metadata in-stream as in the separate file
Previously, the stream contained scalar references to store the only
metadata that was strictly necessary during loading -- the organization.
Switch to using a more extensible hash reference for this, and use the
same structure when storing the "rt-serialized" file which is written as
a summary file after serialization is complete.

We also take this opportunity to begin embedding a format string into
the metadata, as well as the version of RT that was serialized.