load classes then bootstrap DB for tests
If not all classes are loaded then some insert operations
may fail. For example inserting ACL fails when class that
defines a right is not loaded. Many core classes are
already loaded to this moment, but not all. RT::Dashboard
defines a few rights and is not loaded.

We can do this as plugin paths are already set to correct
values and all overrides should be properly loaded. Talking
to DB during loading is prohibitted. If some code tries
to talk to DB while it's loading then it should be changed.