Serializer: Cloning now requires that Principals are serialized
In non-cloning situations, the information from the Principal object of
Users and Groups was serialized onto the latter ofject record, for ease
of dependency analysis, and simplicity when creating Principals.  As
cloning no longer undergoes the custom Serialize and PreInflate methods
that perform this task, cloned serializers must explicitly serialize the
Principals table as well.  While this comes at a cost of space, it
simplifies the logic considerably.