Avoid warnings if CheckIntegrity is called twice as a class method
CheckIntegrity constructs its own RT::Handle object if called as a class
method.  However, this $self not only has an undef DSN, which generates
a warning, but is also misleading, as the DSN that actually failed to
connect is $RT::Handle's.  Consistently use $RT::Handle, as that is what
RT::ConnecttoDatabase sets up, and what RT::CurrentUser looks up in.