Ensure that tests pass no matter if prove is passed -w
7f3d9c1 added the queue name to the warning.  However, in cases where no
Queue parameter is supplied (as in t/web/csrf.t), this can cause
warnings -- but only if the test is run with `prove -w`.  48531b2,
during the merge from 4.0-trunk, modified the test to expect the
warning, but this causes tests to fail when not run with -w.  While
tests should nominally always be run with -w, they should not
mysteriously fail when it is omitted.

Warning-proof the error message so that the test need no longer check
for the (unrelated, in any case) warning.