sync queries in QueueSummaryBy*
put statuses into query, it's already this way in
QueueSummaryByLifecylcle. People rarely limit number of queues
to display, but status is often limited to initial and active.
Intances usually have small fractive of active tickets, so
status condition can be very selective.

Good to go into 4.0 as it just syncs code with QueueSummaryByLifecylcle