Cache CFs in ColumnMap on a per-request basis
When displaying /n/ rows of search results with /m/ columns of global CF
values displayed, custom field queries accounted for /n/ * /m/ * 3
queries; one to try to load the queue-specific CF with the column name,
one to load the global CF with that name, and one to load the OCFV for
that CF and ticket.  With sufficient CFs in a search result format,
these account for a significant performance impact, despite being
relatively well-indexed.

Cache the CF objects in the per-request $m->notes() cache to eliminate
the CF loads, which account for ½ to ⅔ of those queries, depending on
the prevalence of global CFs.  The cache key must also include the
lookup id (queue id, in most cases) -- as a single search may include
results from multiple queues, each of which may resolve a CF name
differently.  It must also include the lookup type, such that a ticket
CF and a transaction CF with the same name can be differentiated.