Remove HomepageSettings preference on "reset" to pick up global defaults
Ensure that the global settings for "RT at a Glance" are taken when a
user resets their preferences.  Previously, the setting was simply set
to an empty hash, which meant that the global defaults were not applied.

This causes the reset button to revert current session to the current
state of the global configuration, which is thereafter cached for the
duration of the session in $session{'my_rt_portlets'}.