Patch timepicker to not force the textbox to update on click-away
Hotpatch the timepicker to remove the part of the onClose method which
forces an unnecessary update from the timepicker.  This allows the user
to type "tomorrow" into the textarea, and have it accepted when the user
clicks clicks away and the datetime picker closes.  This is particularly
important given that the accepted way to remove a due date is to enter
"0" or " " into the input.