Allow altering template type to Simple and changing Content simultaneously
The order in of the array passed to AttributesRef determines the order
in which updates are applied.  Because Content denies updates if the
current Type is 'Perl' and the user doesn't have ExecuteCode, this means
that it was impossible for such a user to update the Content in the same
submission as the type was changed to Simple.  Reorder the AttributesRef
such that Type comes before Content, thus ensuring that the ACL
surrounding Content will be satisfied by the time we attempt to update
it.