Set requestor explicitly, as recipient@example.com is no longer cached
The old 'recipient@example.com' email address was previously incorrectly
cached in the session, despite the user's email address having been
changed to 'general@example.com' in the database.  Now that the
information is reloaded into the session on every request, explicitly
set the requestor on the tickets such that the test continues to pass.