Empty emails should not match empty Correspond addresses on Queues
The value stored in the database for "the default global Correspond
address" is the empty string.  This caused the IsRTAddress code to
false-positive when passed the empty string, such as from a user with no
email address set.  While such addresses should indeed be removed,
catching them in IsRTAddress caused unnecessary warnings in the logs.

Fixes #18380.