change how we deal with decoding structured fields
When we get rid of encoded words in formatted fields (From/To/Cc,
Content-Disposition, Content-Type...), we have to take care of
preserving structure of the field which means quoting parts. Old code
were using luck and magic to achieve this instead of parsing field's
structure.

We use MIME::Field::ParamVal already to deal with continued params
(RFC2231). This module allows us to deal with separate parameters,
decode them and deal with quotes inside or absence of those without
guessing.

Bump the version to a version which fixes a bug [1] in parsing of quoted
strings and quoting on stringify.

Similar situation with From, Cc, Bcc and other fields with mailboxes. We
were only dealing with subset of mailbox fields, now we handle all such
headers that are mentioned in RFC5322.

Switch to Email::Address::List to parse headers with mailboxes. Allows
us to report errors, handles obsolete emails, don't bails out on not
ascii.

Email::Address is still used and a bug [2] was discovered in the module
that may result in unparsable string after re-composing.

[1] https://rt.cpan.org/Ticket/Display.html?id=80433
[2] https://rt.cpan.org/Ticket/Display.html?id=81170