Backport a jQuery UI Mouse patch to fix sliders on IE 9
The bug most obviously affected the hour and minute sliders in the time
picker.

The upstream fix was first released with jQuery UI 1.8.6 as a result of
these three commits: 8fcf58e2, 7eaedaad, ec73a59c.  Patching in-place is
preferred because any version of jQuery UI we update to in the future
will already contain the fix.