We're not cleaning up all CurrentUsers in TransactionBatch
After the sequence of commits around 771a9575a2 we started storing the
CurrentUser of the ticket/transaction in a transaction batch scrip and
then restoring it.  The rationale is in 771a9575a2.  Unfortunately, we
only catch the top-level CurrentUser.  If there are objects cached in
the Ticket or Transaction, they don't get turned into RT_System objects.

This was found because the SLA extension calls
$self->TicketObj->QueueObj->CustomFields and the Unprivileged user
emailing RT doesn't have permission to see the SLA Custom Field.