Encode the interface (Email/Web/Mobile/REST) that content came from
While examining the presence of the Recieved header was often used as a
stand-in for knowing if a transaction was initiated via email or not, it
is simpler to record the interface explicitly.  Store one of four values
in the X-RT-Interface MIME header: Email, Web, Mobile, or REST,
depending on the endpoint of the request that created the MIME object.