Rewrite inline images when serving a single HTML attachment for display
This only happens in the non-default configuration of
AlwaysDownloadAttachments off (default) and TrustHTMLAttachments on
(non-default).  The latter option is primarily used by people so HTML is
rendered in full without the restrictions imposed within ticket history.
It's reasonable to surmise that a site already trusting HTML attachments
for better display will prefer working embedded images.