just call AddCustomField value
First of all calling Delete directly on OCFV is wrong,
doesn't leave a trail and transaction is wrong - says that
value was added, not replaced.

We do report that "thing" was already the current value for
other fields, so I don't see reason to handle it specially
for custom fields.