Process Sign/Encrypt values later on update
A value for Sign or Encrypt of undef is an indication of "do whatever
the queue's options say."  d3341b7 resolved a bug of inserting array
references into X-RT-Sign, but in doing so lost the possibility to pass
undef into ProcessUpdateMessage, and have the queue default kick in.  In
practice, this is only relevant if the signing controls were not
submitted, but customizations may do so.