don't skip conversion from utf8 to utf8
An email may say that it's UTF-8, but it
can contain anything else. We were trusting mail,
but there are too many MUAs, MDAs MTAs that do
things wrong.

If we don't do this conversion then at least on Pg RT fails
to create tickets with error 'invalid byte sequence for encoding
"UTF8"'.