Always throw away warnings caused by compilation tests
e551d98 causes tests to unexpectedly pass because the get_ok inherent to
no_warnings_ok (which is additional to the check that the warnings are
empty) now counts as part of the TODO'd block.

As in this case we do not have any particular expectation that
compilation tests will ever be warnings clean, simply switch to clearing
the caught warnings.