script that deletes bad (Set, TimeWorked) txns
It try to fix history and do it as safe as possible.
It calcs time worked using history, by the way collects
candidates for deletion. Transactions are deleted only
if time on ticket is equal to time from history minus
time on candidates.

It's probably possible to write similar script for TimeLeft
and TimeEstimated, but it's much harder to do as the value
is not recorded on Create unlike time worked which is stored
in TimeTaken column.