Add component and version options to rt-setup-database for upgrade history
Rename realm to component and allow this value to be passed as an
argument to rt-setup-database. This allows extensions to store upgrade
history under their own key.

Add a ext-version argument for extension version and add this as
an optional entry in the upgrade hash.

Module::Install::RTx needs updates to use these new options for
make initdb.