Warn if TransactionBatch would be triggered on shredded tickets
Since 1e280ff, we switch the current user by creating a new RT::Ticket
object and load the current ticket into it.  However, this also has a
subtle behavior change -- previously, if a ticket had ending
transactions and was removed from the underlying database, appropriate
scrips would still be triggered for it, despite the ticket no longer
existing.  Depending on the content of the scrips, this may have led to
unexpected results.  In the new behavior, the ticket object passed to
the scrips unexpectedly has no id.

Before running _ApplyTransactionBatch, detect if the ticket record is no
longer valid, and return, dropping the transactions; in doing so, warn
of the confluence of events that led to this scenario, rare as it is.

While this is technically a behavior change (from both before and after
1e280ff), it almost certainly only arises during testing, and previously
generated subtly incorrect or inconsistent results.  The new behavior
reports the failure, maintains consistency, and suggests a solution.