Update RT::Date RFC2822 to use two-digit month days, fixing RFC 2616 output
While RFC2822 specifies 1*2DIGIT as the 'day' specification [1] (for
which the previous implementation was not incorrect), RFC2616 specifies
2DIGIT [2].  Since the RFC2822 implementation is used in generating the
RFC2616 output, this led to improper date strings when in the first nine
days of each month.

Pad the 'day' portion with 0s as necessary, as required by RFC2616 and
allowed by RFC2822.

Fixes http://issues.bestpractical.com/Ticket/Display.html?id=20701

[1] http://tools.ietf.org/html/rfc2822#section-3.3
[2] http://tools.ietf.org/html/rfc2616#section-3.3.1