Descriptions of just-created transactions should always be readable
If a user had permissions to execute a change, they should always have
permission to read the description of the transaction that is returned.
This case is triggered when a user changes a ticket's queue to one they
do not have rights in, for example; without it, the UI merely displays
the unhelpful status change "Permission denied" in @Actions, despite an
action clearly having happened.

While the simplest method for ensuring that the transaction can be read
would be to load it as the system user, this causes the returned message
to be in the system user's localization, not the current user's.  Add an
explicit flag which allows ->CurrentUserCanSee to be overridden for this
one object.