Remove the unused DeleteCategory method from CFVs
f9afeaf added the code in question, which was correct because, at the
time, Category was an attribute on the CFV.  However, Category has been
a first-class column, and not an attribute, since c0490c3.  Remove the
unnecessary DeleteCategory method, and the Delete method which existed
solely to call it.